Add mockable home fn and `Env` trait
This PR follows up and replaces on #23 which intended to add mockable
functions to the `home` crate. This PR follows in similar vain except
for a few changes.

- All mockable fn and the `Env` trait moved into a new `env` module to
    keep the docs clean from all the extra methods that are usually not
    required for most users of the crate.
- Rename the functions to be `with_env` instead of just having `_from`.

The goal of this PR is to enable `rustup` and hopefully `cargo` in the
future to be able to write tests which mockout the environment while
still going through the core logic of this crate.
